chore(packages): improve the code
