+++ services/simple-staking/src/ui/baby/hooks/api/useSupply.ts
@@ -1,5 +1,6 @@
 import babylon from "@/infrastructure/babylon";
 import { useClientQuery } from "@/ui/common/hooks/client/useClient";
+import { ONE_DAY } from "@/ui/common/constants";
 
 const BABY_SUPPLY_KEY = "BABY_SUPPLY_KEY";
 
@@ -11,5 +12,6 @@ export function useSupply({ enabled = true }: { enabled?: boolean } = {}) {
       return client.baby.getSupply("ubbn");
     },
     enabled,
+    staleTime: ONE_DAY, // Token supply changes very rarely
   });
 }
