+++ services/simple-staking/src/ui/common/constants.ts
@@ -1,5 +1,6 @@
 export const ONE_SECOND = 1000;
 export const ONE_MINUTE = 60 * ONE_SECOND;
+export const ONE_DAY = 24 * 60 * ONE_MINUTE;
 
 export const API_DEFAULT_RETRY_COUNT = 3;
 export const API_DEFAULT_RETRY_DELAY = 3.5; // seconds
